Fix when columns names are reserved SQL keywords
Several functions such as analyze had syntax errors when dealing
with column names that are reserved SQL keywords.
For instance, this is an ugly but valid SQL query:
SELECT `from`.`from` AS `from` FROM `from`.`from`

We improved the quote method to better support this, and make
f.col() automatically quote the column names.